+++ src/background/seelen.rs
@@ -114,6 +114,7 @@ impl Seelen {
 
     pub fn on_settings_change(&mut self) -> Result<()> {
         let state = FULL_STATE.load();
+        rust_i18n::set_locale(&state.locale());
 
         tauri::async_runtime::spawn(async {
             let state = FULL_STATE.load();
@@ -179,13 +180,7 @@ impl Seelen {
         }
     }
 
-    async fn start_async() -> Result<()> {
-        Self::start_ahk_shortcuts().await?;
-        Ok(())
-    }
-
     pub fn start(&mut self) -> Result<()> {
-        SEELEN_IS_RUNNING.store(true, std::sync::atomic::Ordering::SeqCst);
         RestorationAndMigration::run_full()?;
 
         // order is important
@@ -193,6 +188,7 @@ impl Seelen {
         declare_system_events_handlers()?;
 
         let state = FULL_STATE.load();
+        rust_i18n::set_locale(&state.locale());
 
         if state.is_rofi_enabled() {
             self.add_rofi()?;
@@ -215,10 +211,6 @@ impl Seelen {
         MonitorManager::subscribe(Self::on_monitor_event);
         SystemSettings::subscribe(Self::on_system_settings_change);
 
-        tauri::async_runtime::spawn(async {
-            log_error!(Self::start_async().await);
-        });
-
         self.refresh_windows_positions()?;
 
         if FULL_STATE.load().is_weg_enabled() {
@@ -230,6 +222,11 @@ impl Seelen {
         }
 
         register_win_hook()?;
+        tauri::async_runtime::spawn(async {
+            log_error!(Self::start_ahk_shortcuts().await);
+        });
+
+        SEELEN_IS_RUNNING.store(true, std::sync::atomic::Ordering::SeqCst);
         Ok(())
     }