+++ src/background/seelen_wall/mod.rs
@@ -57,7 +57,6 @@ impl SeelenWall {
         .visible(false)
         .disable_drag_drop_handler()
         .skip_taskbar(true)
-        .focused(false)
         // idk why I add this but lively wallpaper has it XD
         // .additional_browser_args("--disk-cache-size=1 --disable-features=msWebOOUI,msPdfOOUI,msSmartScreenProtection")
         .build()?;
@@ -109,9 +108,11 @@ impl SeelenWall {
 
         WindowEnumerator::new().for_each(|current| unsafe {
             // check if current contains SHELLDLL_DefView
-            if FindWindowExA(Some(current), None, pcstr!("SHELLDLL_DefView"), None).is_ok() {
+            if FindWindowExA(Some(current.hwnd()), None, pcstr!("SHELLDLL_DefView"), None).is_ok() {
                 // find next worker after the current one
-                if let Ok(_worker) = FindWindowExA(None, Some(current), pcstr!("WorkerW"), None) {
+                if let Ok(_worker) =
+                    FindWindowExA(None, Some(current.hwnd()), pcstr!("WorkerW"), None)
+                {
                     worker = Some(_worker);
                 }
             }