+++ src/service/main.rs
@@ -10,7 +10,7 @@ mod string_utils;
 mod task_scheduler;
 mod windows_api;
 
-use cli::{handle_console_client, TcpService};
+use cli::{handle_console_client, TcpBgApp, TcpService};
 use crossbeam_channel::{Receiver, Sender};
 use enviroment::was_installed_using_msix;
 use error::Result;
@@ -52,15 +52,6 @@ pub fn stop() {
     STOP_CHANNEL.0.send(()).unwrap();
 }
 
-fn is_seelen_ui_running() -> bool {
-    let mut system = sysinfo::System::new();
-    system.refresh_processes();
-    system
-        .processes()
-        .values()
-        .any(|p| p.exe().is_some_and(|path| path.ends_with("seelen-ui.exe")))
-}
-
 fn launch_seelen_ui() -> Result<()> {
     let app_path = if was_installed_using_msix() {
         WindowsApi::known_folder(FOLDERID_LocalAppData)?
@@ -89,7 +80,7 @@ fn restart_gui_on_crash(max_attempts: u32) {
     std::thread::spawn(move || {
         let mut attempts = 0;
         while attempts < max_attempts {
-            if !is_seelen_ui_running() {
+            if !TcpBgApp::is_running() {
                 attempts += 1;
                 launch_seelen_ui().expect("Failed to launch Seelen UI");
             }
@@ -102,7 +93,7 @@ fn restart_gui_on_crash(max_attempts: u32) {
 #[cfg(debug_assertions)]
 fn stop_service_on_seelen_ui_closed() {
     std::thread::spawn(move || {
-        while is_seelen_ui_running() {
+        while TcpBgApp::is_running() {
             std::thread::sleep(std::time::Duration::from_secs(2));
         }
         stop();
@@ -114,10 +105,12 @@ pub fn setup() -> Result<()> {
     WindowsApi::enable_privilege(SE_TCB_NAME)?;
     TcpService::listen_tcp()?;
 
-    if !is_seelen_ui_running() {
+    if was_started_from_startup_action() {
+        WindowsApi::wait_for_native_shell();
         launch_seelen_ui()?;
     }
 
+    std::thread::sleep(std::time::Duration::from_secs(2));
     #[cfg(debug_assertions)]
     {
         stop_service_on_seelen_ui_closed();
@@ -152,7 +145,7 @@ fn main() -> Result<()> {
         return Ok(());
     }
 
-    SluServiceLogger::install()?;
+    let _ = SluServiceLogger::uninstall_old_logging();
     SluServiceLogger::init()?;
     TaskSchedulerHelper::create_service_task()?;